Moved window functions into .c file.
